+++ FeedStoreChallenge/CodableFeedStore.swift
@@ -1,0 +1,93 @@
+//
+//  CodableFeedStore.swift
+//  FeedStoreChallenge
+//
+//  Created by Raphael Silva on 30/12/2019.
+//  Copyright © 2019 Essential Developer. All rights reserved.
+//
+
+import Foundation
+
+public class CodableFeedStore: FeedStore {
+    private struct Cache: Codable {
+        let feed: [CodableFeedImage]
+        let timestamp: Date
+        
+        var localFeed: [LocalFeedImage] {
+            return feed.map { $0.local }
+        }
+    }
+    
+    private struct CodableFeedImage: Codable {
+        private let id: UUID
+        private let description: String?
+        private let location: String?
+        private let url: URL
+        
+        var local: LocalFeedImage {
+            return LocalFeedImage(id: id, description: description, location: location, url: url)
+        }
+        
+        init(_ localFeedImage: LocalFeedImage) {
+            id = localFeedImage.id
+            description = localFeedImage.description
+            location = localFeedImage.location
+            url = localFeedImage.url
+        }
+    }
+    
+    private let storeURL: URL
+    private let queue = DispatchQueue(label: "\(CodableFeedStore.self)Queue", qos: .default, attributes: .concurrent)
+    
+    public init(storeURL: URL) {
+        self.storeURL = storeURL
+    }
+    
+    public func deleteCachedFeed(completion: @escaping DeletionCompletion) {
+        let storeURL = self.storeURL
+        queue.async(flags: .barrier) {
+            let fileManager = FileManager.default
+            guard fileManager.fileExists(atPath: storeURL.path) else {
+                return completion(nil)
+            }
+            
+            do {
+                try fileManager.removeItem(at: storeURL)
+                completion(nil)
+            } catch {
+                completion(error)
+            }
+        }
+    }
+    
+    public func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
+        let storeURL = self.storeURL
+        queue.async(flags: .barrier) {
+            do {
+                let encoder = JSONEncoder()
+                let encodedData = try encoder.encode(Cache(feed: feed.map(CodableFeedImage.init), timestamp: timestamp))
+                try encodedData.write(to: storeURL)
+                completion(nil)
+            } catch {
+                completion(error)
+            }
+        }
+    }
+    
+    public func retrieve(completion: @escaping RetrievalCompletion) {
+        let storeURL = self.storeURL
+        queue.async {
+            guard let storedData = try? Data(contentsOf: storeURL) else {
+                return completion(.empty)
+            }
+            
+            do {
+                let decoder = JSONDecoder()
+                let cache = try decoder.decode(Cache.self, from: storedData)
+                completion(.found(feed: cache.localFeed, timestamp: cache.timestamp))
+            } catch {
+                completion(.failure(error))
+            }
+        }
+    }
+}
